requirements.txt
coqui-tts
ffmpeg
openai-whisper
tqdm
